yuuuu

+++ Assets/Script/Events/CharacterEvents.cs
@@ -5,10 +5,10 @@ using UnityEngine.Events;
 
 public class CharacterEvents
 {
-    public static UnityAction<GameObject, int> characterDamaged;
+    public static UnityAction<GameObject, float> characterDamaged;
 
 
-    public static UnityAction<GameObject, int> characterHealed;
+    public static UnityAction<GameObject, float> characterHealed;
 
 
 }
